This *should* fix kernel cache stuff for svmlight:
+kernel cache indices are now checked to fit in KERNELCACHE_IDX type 
	an assertion is thrown if not
+kernel cache is reduced to totdoc*totdoc*sizeof(KERNELCACHE_ELEM) if larger
+KERNELCACHE_IDX is now LONG again, access operations to the cache buffer do a 
	cast to KERNELCACHE_IDX, while the cache lines are limited to INTs only 
	(i.e. caching max ~2^31 cache lines (~2^62 elements) is possible)

+++ src/lib/common.h
@@ -95,7 +95,7 @@ typedef LONGREAL* P_LONGREAL;
 	typedef DREAL KERNELCACHE_ELEM;
 #endif
 
-typedef INT KERNELCACHE_IDX;
+typedef LONG KERNELCACHE_IDX;
 
 /// The io libs output [DEBUG] etc in front of every CIO::message
 /// 'higher' messages filter output depending on the loglevel, i.e. CRITICAL messages
